# Optimize send response