--- a/core/include/userver/server/http/http_response.hpp
+++ b/core/include/userver/server/http/http_response.hpp
@@ -15,6 +15,7 @@
 #include <userver/server/request/response_base.hpp>
 #include <userver/utils/impl/projecting_view.hpp>
 #include <userver/utils/str_icase.hpp>
+#include <userver/utils/small_string.hpp>
 
 #include "http_status.hpp"
 
@@ -139,11 +140,13 @@ class HttpResponse final : public request::ResponseBase {
 
  private:
   // Returns total size of the response
-  std::size_t SetBodyStreamed(engine::io::RwBase& socket, std::string& header);
+  template <std::size_t N>
+  std::size_t SetBodyStreamed(engine::io::RwBase& socket, utils::SmallString<N>& header);
 
   // Returns total size of the response
+  template <std::size_t N>
   std::size_t SetBodyNotStreamed(engine::io::RwBase& socket,
-                                 std::string& header);
+                                 utils::SmallString<N>& header);
 
   const HttpRequestImpl& request_;
   HttpStatus status_ = HttpStatus::kOk;