--- a/core/include/userver/server/http/http_response.hpp
+++ b/core/include/userver/server/http/http_response.hpp
@@ -14,8 +14,8 @@
 #include <userver/server/http/http_response_cookie.hpp>
 #include <userver/server/request/response_base.hpp>
 #include <userver/utils/impl/projecting_view.hpp>
-#include <userver/utils/str_icase.hpp>
 #include <userver/utils/small_string.hpp>
+#include <userver/utils/str_icase.hpp>
 
 #include "http_status.hpp"
 
@@ -139,14 +139,16 @@ class HttpResponse final : public request::ResponseBase {
   Queue::Producer GetBodyProducer();
 
  private:
+  static constexpr std::size_t kTypicalHeaderSize = 1024;
+
+  using HeadersString = utils::SmallString<kTypicalHeaderSize>;
   // Returns total size of the response
-  template <std::size_t N>
-  std::size_t SetBodyStreamed(engine::io::RwBase& socket, utils::SmallString<N>& header);
+  std::size_t SetBodyStreamed(engine::io::RwBase& socket,
+                              HeadersString& header);
 
   // Returns total size of the response
-  template <std::size_t N>
   std::size_t SetBodyNotStreamed(engine::io::RwBase& socket,
-                                 utils::SmallString<N>& header);
+                                 HeadersString& header);
 
   const HttpRequestImpl& request_;
   HttpStatus status_ = HttpStatus::kOk;