--- a/core/include/userver/server/http/http_response.hpp
+++ b/core/include/userver/server/http/http_response.hpp
@@ -24,6 +24,9 @@ namespace server::http {
 
 namespace impl {
 
+void OutputHeader(std::string& header, std::string_view key,
+                  std::string_view val);
+
 void OutputHeader(USERVER_NAMESPACE::http::headers::HeadersString& header,
                   std::string_view key, std::string_view val);