--- a/core/include/userver/server/http/http_response_cookie.hpp
+++ b/core/include/userver/server/http/http_response_cookie.hpp
@@ -9,8 +9,8 @@
 #include <string>
 #include <unordered_map>
 
-#include <userver/utils/small_string.hpp>
 #include <userver/utils/str_icase.hpp>
+#include <userver/http/predefined_header.hpp>
 
 USERVER_NAMESPACE_BEGIN
 
@@ -64,8 +64,7 @@ class Cookie final {
   std::string ToString() const;
 
   void AppendToString(std::string& os) const;
-  template <std::size_t N>
-  void AppendToString(utils::SmallString<N>& os) const;
+  void AppendToString(userver::http::headers::HeadersString& os) const;
 
  private:
   class CookieData;