--- a/core/include/userver/server/http/http_response_cookie.hpp
+++ b/core/include/userver/server/http/http_response_cookie.hpp
@@ -64,7 +64,6 @@ class Cookie final {
   USERVER_NAMESPACE::http::headers::HeadersString ToSmallString() const;
   std::string ToString() const;
 
-  void AppendToString(std::string& os) const;
   void AppendToString(USERVER_NAMESPACE::http::headers::HeadersString& os) const;
 
  private: