--- a/core/include/userver/server/http/http_response_cookie.hpp
+++ b/core/include/userver/server/http/http_response_cookie.hpp
@@ -61,7 +61,7 @@ class Cookie final {
   std::string SameSite() const;
   Cookie& SetSameSite(std::string value);
 
-  USERVER_NAMESPACE::http::headers::HeadersString ToString() const;
+  std::string ToString() const;
 
   void AppendToString(
       USERVER_NAMESPACE::http::headers::HeadersString& os) const;