Оптимизация HttpResponse::SendResponse с помощью замены std::string header на userver::utils::SmallString (#407)
```
---------------------------------------------------------------------------------
Benchmark                                       Time             CPU   Iterations
---------------------------------------------------------------------------------
http_headers_serialization_inplace           72.5 ns         72.4 ns      9808811
http_headers_serialization_no_ostreams        264 ns          264 ns      2481104
http_headers_serialization_ostreams           755 ns          755 ns       923180

```